requirements.txt
python-dotenv 
google-genai
google-generativeai
streamlit
unstructured
unstructured[pdf]
langchain==0.3.13
langchain_community==0.3.13
sentence_transformers==3.3.1
nltk
tf-keras
python-magic
faiss-cpu
rank_bm25